update dToken for gas optimisations and extra events

+++ protocol/contracts/fakes/v2/TempleDebtTokenTestnetAdmin.sol
@@ -32,7 +32,7 @@ contract TempleDebtTokenTestnetAdmin {
         dUSD.setBaseInterestRate(_rate);
     }
 
-    function setRiskPremiumInterestRate(address _debtor, uint64 _rate) external {
+    function setRiskPremiumInterestRate(address _debtor, uint96 _rate) external {
         dUSD.setRiskPremiumInterestRate(_debtor, _rate);
     }
